mpt ~Node uint8_t->unsigned for loop comparison

+++ libs/db/src/monad/mpt/node.cpp
@@ -53,9 +53,8 @@ Node::Node(
     NibblesView const path, int64_t const version)
     : mask(mask)
     , path_nibble_index_end(path.end_nibble_)
-    , value_len(
-          static_cast<decltype(value_len)>(
-              value.transform(&byte_string_view::size).value_or(0)))
+    , value_len(static_cast<decltype(value_len)>(
+          value.transform(&byte_string_view::size).value_or(0)))
     , version(version)
 {
     MONAD_DEBUG_ASSERT(
@@ -80,7 +79,7 @@ Node::Node(
 
 Node::~Node()
 {
-    for (uint8_t index = 0; index < number_of_children(); ++index) {
+    for (unsigned index = 0; index < number_of_children(); ++index) {
         move_next(index).reset();
     }
 }
